+++ src/App/components/IconPreview/index.js
@@ -2,31 +2,24 @@ import React from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
 
-const PaymentIcon = ({ name, preview, previewSize, size, className }) => {
+const IconPreview = ({ name, preview, previewSize, size, className, type }) => {
     const classNames = className ? className.split(" ") : [];
-    const iconClasses = classnames("payment-icon", name, size, ...classNames);
+    const iconClasses = classnames(type, type === "payment-icon" ? name : "", size, ...classNames);
 
     return preview || previewSize ? (
         <div className="icon-preview">
-            <i className={iconClasses}></i>
+            <i className={iconClasses}>{iconClasses.includes("material-icons") ? name : null}</i>
             <span>{preview ? name : size}</span>
         </div>
     ) : <i className={iconClasses}></i>;
 };
 
-PaymentIcon.propTypes = {
-    name: PropTypes.oneOf([
-        "amex",
-        "diners",
-        "mastercard",
-        "mobilepay",
-        "paypal",
-        "swish",
-        "vipps",
-        "visa"
-    ]).isRequired,
+IconPreview.propTypes = {
+    name: PropTypes.string.isRequired,
+    type: PropTypes.string.isRequired,
     preview: PropTypes.bool,
     size: PropTypes.oneOf([
+        "tiny",
         "small",
         "medium",
         "large",
@@ -35,4 +28,4 @@ PaymentIcon.propTypes = {
     className: PropTypes.string
 };
 
-export default PaymentIcon;
+export default IconPreview;