+++ CHANGELOG.md
@@ -6,9 +6,14 @@
 
 - Added documentation for image utilities.
 - Added image utility .img-thumbnail.
+- Added size classes for material icons.
 
 ### Changed
+
+- Renamed icons to iconography under core, and moved payment icons under material icons.
 
 ### Fixed
 
 ### Removed
+
+- Removed Payment Icons from core documentation.